major refactoring - writeup to follow...

+++ bcel-builder/src/org/aspectj/apache/bcel/classfile/ConstantNameAndType.java
@@ -63,7 +63,7 @@ import  java.io.*;
  * and represents a reference to the name and signature
  * of a field or method.
  *
- * @version $Id: ConstantNameAndType.java,v 1.2 2004/11/19 16:45:18 aclement Exp $
+ * @version $Id: ConstantNameAndType.java,v 1.2.10.1 2007/02/12 09:34:02 aclement Exp $
  * @author  <A HREF="mailto:markus.dahm@berlin.de">M. Dahm</A>
  * @see     Constant
  */
@@ -86,7 +86,7 @@ public final class ConstantNameAndType extends Constant {
    */
   ConstantNameAndType(DataInputStream file) throws IOException
   {    
-    this((int)file.readUnsignedShort(), (int)file.readUnsignedShort());
+    this(file.readUnsignedShort(), file.readUnsignedShort());
   }
 
   /**
